tidied up some warning messages
git-svn-id: https://svn.code.sf.net/p/pianobooster/code/trunk@47 ba081f5d-443b-49a7-ac4b-446c3f91f371

+++ PianoBooster/src/Draw.cpp
@@ -452,7 +452,7 @@ void CDraw::drawSymbol(CSymbol symbol, float x, float y)
     if (symbol.getType() == PB_SYMBOL_note)
     {
         drawStaveExtentsion(symbol, x, 16);
-        int pianistX = symbol.getPianistTiming();
+        float pianistX = symbol.getPianistTiming();
         if ( pianistX != NOT_USED)
         {
             pianistX =  x + pianistX * HORIZONTAL_SPACING_FACTOR;
